requirements.txt
django
psycopg2-binary
dnspython
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-environ
requests
pdfminer.six
boto3
langchain
langchain_aws